changed footer to be horizontally
Imp: I had to add classes in extra divs where there was only ids to be easier to manipulate (this could be the issue to center without flex)
Issues: The 3 divs inside still needs to be centered. Main seems to be too much to the right, needs to be checked if it was like that before.
@AMJaneiro